Some changes are made

+++ Nocturnal/source/globals.cpp
@@ -15,7 +15,7 @@ Vector2 playerPos = GetMousePosition();
 float radius = 10.0f;
 
 // Enemy variables
-Vector2 enemyPos = { GetRandomValue(10, windowWidth - 100), GetRandomValue(10, windowHeight - 100) };
+Vector2 enemyPos = { static_cast<float>(GetRandomValue(10, windowWidth - 100)), static_cast<float>(GetRandomValue(10, windowHeight - 100)) };
 Rectangle enemyRec = { enemyPos.x, enemyPos.y, 50, 50 };
 float spawnTime = 0.0f;
 bool isAlive = true;
